Move email policy to server email settings
