+++ BTCPayServer/Models/ServerViewModels/EmailsViewModel.cs
@@ -1,32 +1,18 @@
 using System.ComponentModel.DataAnnotations;
 using BTCPayServer.Services.Mails;
-using BTCPayServer.Validation;
 
-namespace BTCPayServer.Models.ServerViewModels
+namespace BTCPayServer.Models.ServerViewModels;
+
+public class ServerEmailsViewModel : EmailsViewModel
 {
-    public class EmailsViewModel
+    [Display(Name = "Allow stores to use the server's SMTP email settings as a default")]
+    public bool EnableStoresToUseServerEmailSettings { get; set; }
+
+    public ServerEmailsViewModel()
     {
-        public EmailSettings Settings { get; set; }
-        public EmailSettings FallbackSettings { get; set; }
-        public bool PasswordSet { get; set; }
-        
-        [MailboxAddress]
-        [Display(Name = "Test Email")]
-        public string TestEmail { get; set; }
+    }
 
-        public EmailsViewModel()
-        {
-        }
-
-        public EmailsViewModel(EmailSettings settings, EmailSettings fallbackSettings = null)
-        {
-            Settings = settings;
-            FallbackSettings = fallbackSettings;
-            PasswordSet = !string.IsNullOrEmpty(settings?.Password);
-        }
-        
-        public bool IsSetup() => Settings?.IsComplete() is true;
-        public bool IsFallbackSetup() => FallbackSettings?.IsComplete() is true;
-        public bool UsesFallback() => IsFallbackSetup() && Settings == FallbackSettings;
+    public ServerEmailsViewModel(EmailSettings settings) : base(settings)
+    {
     }
 }
